renamed original to full_state_space
