+++ doc/full_state_space/PerformanceCriticalCode.rst
@@ -1,9 +1,9 @@
 :orphan:
 
-The *RegularizedOptimization* submodule
+The *PerformanceCriticalCode* submodule
 =======================================
 
-.. automodule:: mhn.original.RegularizedOptimization
+.. automodule:: mhn.full_state_space.PerformanceCriticalCode
    :members:
    :undoc-members:
    :show-inheritance:
