+++ doc/full_state_space/UtilityFunctions.rst
@@ -3,7 +3,7 @@
 The *UtilityFunctions* submodule
 ================================
 
-.. automodule:: mhn.original.UtilityFunctions
+.. automodule:: mhn.full_state_space.UtilityFunctions
    :members:
    :undoc-members:
    :show-inheritance:
